feat: sync common rules interfaces (#41)
* feat: sync IDL for `parentRule` `parentStyleSheet` `selectorText` `style`

CSSRule `parentRule` and `parentStyleSheet`, are now prototype getter to be readonly according to the CSSRule interface.
CSSStyleRule `selectorText` and `style` are now prototype getter and setter according to the CSSStyleRule interface.

* feat: use `csstyle` package when available

To be used by jsdom

* feat: print the getters back in docs inspect

* feat: hide the __ prefixed props in docs inspect

* feat: add missing IDL for `style` on rules

* feat: implement errorUtils

* feat: implement CSSKeyframesRule methods

appendRule, deleteRule, findRule and indexed getters

* fix: adjust enumerable to false on indexed getters

* fix: update docs and spec

* chore: build fix

* chore: add changeset

Updated CSSRule and CSSStyleRule to align with their interfaces and improved error handling.